fix(switcher): decode fractional cent values in usage/balance
reseller_usage_logs.cost_cents and reseller_api_keys.spent_cents/
budget_cents are Postgres numeric columns. Since per-model pricing
landed they carry sub-cent fractions (e.g. 4.0573), which serialize as
JSON floats. The Tauri structs declared these fields i64, so serde_json
rejected the float ('invalid type: floating point, expected i64') and
reqwest surfaced it as 'error decoding response body' -- blanking the
Usage page. Widen cost_cents/spent_cents/balance_cents to f64.

Verified against prod data: old i64 struct fails, new f64 struct decodes.

+++ grouter-switcher/src-tauri/src/commands.rs
@@ -117,10 +117,12 @@ pub async fn recover_account(state: State<'_, AppState>, recovery_password: Stri
 #[derive(Serialize)]
 pub struct BalanceResult {
     pub username: String,
+    // balance/spent are derived from `numeric` budget_cents/spent_cents and can
+    // be fractional, so keep them as f64 (see UsageEntry::cost_cents).
     #[serde(rename = "balanceCents")]
-    pub balance_cents: Option<i64>,
+    pub balance_cents: Option<f64>,
     #[serde(rename = "spentCents")]
-    pub spent_cents: i64,
+    pub spent_cents: f64,
     pub unlimited: bool,
 }
 
@@ -128,9 +130,9 @@ pub struct BalanceResult {
 struct MeResponseBody {
     username: String,
     #[serde(rename = "balanceCents")]
-    balance_cents: Option<i64>,
+    balance_cents: Option<f64>,
     #[serde(rename = "spentCents")]
-    spent_cents: i64,
+    spent_cents: f64,
     unlimited: bool,
 }
 
@@ -173,8 +175,12 @@ pub struct UsageEntry {
     pub cache_creation_input_tokens: i64,
     #[serde(rename = "cacheReadInputTokens")]
     pub cache_read_input_tokens: i64,
+    // cost_cents is a Postgres `numeric` and, since per-model pricing landed,
+    // carries sub-cent fractions (e.g. 4.0573). It serializes as a JSON float,
+    // so this must be f64 -- serde_json refuses to decode a float into an int
+    // and reqwest surfaced that as "error decoding response body".
     #[serde(rename = "costCents")]
-    pub cost_cents: i64,
+    pub cost_cents: f64,
     pub stream: bool,
     #[serde(rename = "createdAt")]
     pub created_at: String,
@@ -184,8 +190,9 @@ pub struct UsageEntry {
 pub struct UsageResult {
     #[serde(rename = "requestCount")]
     pub request_count: i64,
+    // Fractional (numeric) -- see UsageEntry::cost_cents.
     #[serde(rename = "costCents")]
-    pub cost_cents: i64,
+    pub cost_cents: f64,
     #[serde(rename = "inputTokens")]
     pub input_tokens: i64,
     #[serde(rename = "outputTokens")]
